worker: Split action library into its own module
Change-type: patch
