+++ lib/action-server.js
@@ -15,7 +15,7 @@
  */
 
 const Worker = require('./worker')
-const library = require('./worker/library')
+const actionLibrary = require('./action-library')
 
 const SCHEMA_ACTIVE_TRIGGERS = {
 	type: 'object',
@@ -78,7 +78,7 @@ const processQueue = (jellyfish, worker, session, onRequest, onError) => {
 }
 
 exports.start = async (jellyfish, session, onRequest, onError) => {
-	const worker = new Worker(jellyfish, session, library)
+	const worker = new Worker(jellyfish, session, actionLibrary)
 	const triggerStream = await jellyfish.stream(session, SCHEMA_ACTIVE_TRIGGERS)
 	triggerStream.on('error', onError)
